plugins-core/agent-arbitrage-agent/src/requirements.txt
websocket-client
requests
python-dotenv
py-clob-client
web3
